+++ lib/src/solid/utils/large_file_helper.dart
@@ -80,6 +80,31 @@ Future<void> readLargeFile({
   await sink.close();
 }
 
+/// Get a large file previously sent using [writeLargeFile] with name
+/// [remoteFileName] and return it as bytes.
+Future<Uint8List> readLargeFileAsBytes({
+  required String remoteFileName,
+  required BuildContext context,
+  required Widget child,
+  String? ownerWebId,
+  void Function(int, int)? onProgress,
+}) async {
+  final chunks = fetch(
+    remoteFileName: remoteFileName,
+    context: context,
+    child: child,
+    ownerWebId: ownerWebId,
+    onProgress: onProgress,
+  );
+  final builder = BytesBuilder();
+
+  await for (final chunk in chunks) {
+    builder.add(chunk);
+  }
+
+  return builder.toBytes();
+}
+
 /// Send a large local file with path [localFilePath] to a remote server
 /// using name [remoteFileName],
 /// encrypt the file content if [encrypted] is true.