+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Visit the package at [pub.dev](https://pub.dev/packages/solidpod).
 
 ## 0.9 Stabilise
 
++ Enable in-app shared file reload [0.8.7 20251211 anushkavid]
++ Updated dependencies [0.8.6 20251206 gjw]
++ Function to read large file as bytes [0.8.5 20251201 anushkavid]
 + Add PathType to readPod/WritePod [0.8.4 20251123 cdawei]
 + Cleanup and update [0.8.3 20251121 gjw]
 + Allow concurrent pod resource reading [0.8.2 20251117 jesscmoore]